Apologies for the weird single commit—this was an old project.
Back then, I wasn’t really familiar with committing or using GitHub properly.
The structure isn’t the best, but it works.

This was just a beginner project I made to learn the basics,
and I didn’t want to just delete it.

It’s a nice memory, and I’m keeping it as a reminder of how I started. :D